% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huagrahuma.dem}
\alias{huagrahuma.dem}
\title{DEM of the Huagrahuma catchment, Ecuador}
\format{
A matrix representing a digital elevation model
}
\source{
Digital Elevation Model of the Huagrahuma microcatchment, Ecuador, at 25 m resolution
}
\usage{
huagrahuma.dem
}
\description{
Matrix containing a Digital Elevation Model of the Huagrahuma microcatchment, Ecuador, at 25 m resolution.
}
\references{
Buytaert, W., De Bievre, B., Wyseure, G., Deckers, J., 2005. The effect of land use changes on the hydrological
behaviour of Histic Andosols in south Ecuador. Hydrological Processes 19: 3985 - 3997.
}
\keyword{datasets}
