% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSeff.R
\name{NSeff}
\alias{NSeff}
\title{Calculation of the Nash-Sutcliffe efficiency}
\usage{
NSeff(Qobs, Qsim)
}
\arguments{
\item{Qobs}{A vector with observed discharges}

\item{Qsim}{A vector with simulated discharges of the same length as Qobs}
}
\description{
Calculation of the Nash-Sutcliffe efficiency
}
\details{
Qobs and Qsim should have the same dimensions. Both can contain NA values.
}
\seealso{
\code{\link{topmodel}}
}
\author{
Wouter Buytaert, Imperial College London
}
