% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcatch.R
\name{subcatch}
\alias{subcatch}
\title{Identify a hydrological catchment based on a single direction flow algorithm}
\usage{
subcatch(DEM, outlet)
}
\arguments{
\item{DEM}{A matrix representing a digital elevation model [m] with equally sized pixels and equal x and y resolution}

\item{outlet}{A vector containing the row and column indices of the pixel
representing the catchment outlet.}
}
\value{
A binary matrix of the same size as \code{DEM}.
}
\description{
Identify a hydrological catchment based on a single direction flow algorithm
}
\references{
See \url{https://github.com/ICHydro/topmodel} for examples.
}
\seealso{
\code{\link{topmodel}}
}
\author{
Wouter Buytaert, Imperial College London, based on an implementation from the
Hydrology Group of Lancaster University
}
