% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topmodel-package.R
\docType{package}
\name{topmodel-package}
\alias{topmodel-package}
\title{topmodel: Implementation of the Hydrological Model TOPMODEL in R}
\description{
Set of hydrological functions including an R implementation of the hydrological model TOPMODEL, which is based on the 1995 FORTRAN version by Keith Beven. From version 0.7.0, the package is put into maintenance mode.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ICHydro/topmodel}
}

}
\keyword{internal}
