\name{huagrahuma.dem}
\docType{data}
\alias{huagrahuma.dem}
\title{DEM of the Huagrahuma catchment, Ecuador}
\description{
  Matrix containing a Digital Elevation Model of the Huagrahuma microcatchment, Ecuador, at 25 m resolution.
}
\usage{huagrahuma.dem}
%\format{}
%\source{}
\references{
  Buytaert, W., De Bievre, B., Wyseure, G., Deckers, J., 2005. The effect of land use changes on the hydrological
  behaviour of Histic Andosols in south Ecuador. Hydrological Processes 19: 3985 - 3997.
}
\keyword{datasets}

