\name{huagrahuma}
\docType{data}
\alias{huagrahuma}
\alias{delay}
\alias{ETp}
\alias{parameters}
\alias{rain}
\alias{Qobs}
\title{Hydrological dataset to run TOPMODEL}
\description{
  This is an example dataset from a small mountainous catchment in the South American Andes and contains the following data:

  \tabular{ll}{
    delay      \tab Cumulative delay function of the catchment\cr
    ETp        \tab FAO Penman Monteith potential evapotranspiration [m / 15 min]\cr
    parameters \tab Values for the TOPMODEL parameters\cr
    Qobs       \tab Observed discharge [m / 15 min]\cr
    rain       \tab Observed precipitation [m / 15 min]\cr
  }
}
\usage{data(huagrahuma)}
%\format{}
%\source{}
\references{
  Buytaert, W., De Bievre, B., Wyseure, G., Deckers, J., 2005. The effect of land use changes on the hydrological
  behaviour of Histic Andosols in south Ecuador. Hydrological Processes 19: 3985 - 3997.
}
\keyword{datasets}

