\name{NSeff}
\alias{NSeff}
\title{Nash-Sutcliffe efficiency}
\description{Calculation of the Nash-Sutcliffe efficiency between an observed and simulated discharge series}
\usage{NSeff(Qobs,Qsim)}

\arguments{
  \item{Qobs}{A vector with observed discharges}
  \item{Qsim}{A vector with simulated discharges of the same length as Qobs}
}

\details{
  Qobs and Qsim should have the same dimensions. Both can contain NA values.
}
\value{}
% \references{}
\author{Wouter Buytaert, University of Bristol}
\note{}

% \seealso{\code{\link{topmodel.topmodel}}}
\examples{}
\keyword{models}
