\name{NSeff}
\alias{NSeff}
\title{Calculation of the Nash-Sutcliffe efficiency}
\description{Calculation of the Nash-Sutcliffe efficiency}
\usage{NSeff(Qobs,Qsim)}

\arguments{
  \item{Qobs}{A vector with observed discharges}
  \item{Qsim}{A vector with simulated discharges of the same length as Qobs}
}

\details{
  Qobs and Qsim should have the same dimensions. Both can contain NA values.
}
%\value{}
% \references{}
\author{Wouter Buytaert, Imperial College London}
%\note{}

\seealso{\code{\link{topmodel}}}
%\examples{}
\keyword{models}
